requirements.txt
anyio==4.13.0
asgiref==3.11.1
black==26.3.1
certifi==2026.4.22
click==8.3.2
Django==6.0.4
django-rest-framework==0.1.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.0.4
platformdirs==4.9.6
PyJWT==2.12.1
python-dotenv==1.2.2
python-telegram-bot==22.7
pytokens==0.4.1
sqlparse==0.5.5
typing_extensions==4.15.0

requests